MINOR: Move DeleteGroupsResult to internals package.
